requirements.txt
pillow
numpy
grpc
